Rename dateRange module to more generic date
